Incorporation of QAudioCoder. CMake updates for 2.8.8 and Qt5. New Qwt trunk checkout. Lame, Flac and zlib headers added.

+++ code/external/qwt/headers/qwt_round_scale_draw.h
@@ -35,8 +35,8 @@ public:
     QwtRoundScaleDraw();
     virtual ~QwtRoundScaleDraw();
 
-    void setRadius( int radius );
-    int radius() const;
+    void setRadius( double radius );
+    double radius() const;
 
     void moveCenter( double x, double y );
     void moveCenter( const QPointF & );
